pyproject.toml
[project]
name = "auto-heroic-categories"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4==4.12.2",
    "certifi==2024.2.2",
    "charset-normalizer==3.3.2",
    "distlib==0.3.8",
    "filelock==3.14.0",
    "idna==3.7",
    "pipenv==2023.12.1",
    "platformdirs==4.2.1",
    "protobuf==5.26.1",
    "python-dotenv==1.0.1",
    "requests==2.31.0",
    "setuptools==69.5.1",
    "urllib3==2.2.1",
    "virtualenv==20.26.2",
]
